Adding spaces and moving the Converter class to common
